pyproject.toml
[project]
name = "linuxdo-checkin"
version = "2.0.0"
description = "多平台签到工具 - 支持 LinuxDo 论坛和 AnyRouter/AgentRouter API 中转站签到"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "linuxdo-checkin contributors" }
]
keywords = ["checkin", "linuxdo", "anyrouter", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Browser automation
    # Camoufox = anti-detection Firefox (recommended for Cloudflare bypass)
    "camoufox[geoip]>=0.4.0",
    # Patchright = anti-detection Playwright (fallback)
    "patchright>=1.49.0",
    # HTTP clients
    "httpx[http2]>=0.25.0",
    "curl-cffi>=0.5.0",
    # Logging
    "loguru>=0.7.2",
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    # Utilities
    "wcwidth>=0.2.13",
    "tabulate>=0.9.0",
    "customtkinter>=5.2.2",
    "browser-cookie3>=0.20.1",
    "rookiepy>=0.5.0",
    "drissionpage>=4.1.1.2",
    "nodriver>=0.48.1",
    "leveldb-py>=0.0.1b0",
    "pillow>=12.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "pytest-cov>=4.1.0",
    
    # Code quality
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
linuxdo-checkin = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["platforms", "utils"]
include = ["main.py"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = [
    ".git",
    ".github",
    ".kiro",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
    "ARG001", # unused function argument (common in adapters)
]

[tool.ruff.lint.isort]
known-first-party = ["platforms", "utils"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    ".git",
    ".github",
    ".kiro",
    ".venv",
]

[tool.coverage.run]
source = ["platforms", "utils"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
